added unit tests

+++ src/main/kotlin/project/vendingmachine/models/ChangeRequest.kt
@@ -1,5 +1,6 @@
 package project.vendingmachine.models
 
-class ChangeRequest(var amount: Int) {
+
+class ChangeRequest(var amount: Double) {
 
 }
